+++ pages/changelog/index.tsx
@@ -16,7 +16,7 @@ import {
 } from '../../components/Changelog/ChangelogEntry/ChangelogEntry';
 import { GetStaticProps } from 'next/types';
 import { Meta } from '../../components/common/Head/Meta';
-import { GraphQLRequest } from '../util';
+import { GraphQLRequest } from '../../utils/graphql';
 
 const ITEMS_PER_PAGE = 25;
